Backend/requirements.txt
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.0
django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==24.1.1
h11==0.16.0
packaging==26.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyjwt==2.10.1
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
uvicorn==0.40.0
whitenoise==6.11.0
django-phonenumber-field[phonenumbers]==8.0.0
Pillow==11.1.0